Move overlay related components to new "Reusable UI" directory
