+++ ui/src/flux/components/FluxHeader.tsx
@@ -1,7 +1,7 @@
 import React, {PureComponent} from 'react'
 
 import FluxOverlay from 'src/flux/components/FluxOverlay'
-import OverlayTechnology from 'src/shared/components/OverlayTechnology'
+import OverlayTechnology from 'src/reusable_ui/components/overlays/OverlayTechnology'
 import PageHeader from 'src/shared/components/PageHeader'
 
 import {Service} from 'src/types'
